packages/phlo-nessie/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phlo-nessie"
version = "0.1.0"
description = "Nessie service plugin for Phlo"
readme = {text = "Nessie service plugin for Phlo.", content-type = "text/plain"}
requires-python = ">=3.11"
authors = [
    {name = "Phlo Team", email = "team@phlo.dev"},
]
license = {text = "MIT"}
dependencies = [
    "phlo>=0.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.entry-points."phlo.plugins.services"]
nessie = "phlo_nessie.plugin:NessieServicePlugin"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
phlo_nessie = ["service.yaml"]

[tool.ruff]
line-length = 100
target-version = "py311"
